moved everything into interfaces
